Fix an obvious bug in utility.h

+++ src/utility.h
@@ -222,13 +222,13 @@ inline u16 readU16(std::istream &is)
 	return readU16((u8*)buf);
 }
 
-inline void writeU32(std::ostream &os, u16 p)
+inline void writeU32(std::ostream &os, u32 p)
 {
 	char buf[4];
 	writeU16((u8*)buf, p);
 	os.write(buf, 4);
 }
-inline u16 readU32(std::istream &is)
+inline u32 readU32(std::istream &is)
 {
 	char buf[4];
 	is.read(buf, 4);
